Refactor local filter and add keyword filter
