+++ src/lib/Scenes/MyCollection/MyCollection.tsx
@@ -1,15 +1,10 @@
 import { addCollectedArtwork, OwnerType } from "@artsy/cohesion"
 import AsyncStorage from "@react-native-community/async-storage"
+import { InfiniteScrollArtworksGrid_myCollectionConnection } from "__generated__/InfiniteScrollArtworksGrid_myCollectionConnection.graphql"
 import { MyCollection_me } from "__generated__/MyCollection_me.graphql"
-import { MyCollectionArtworkListItem_artwork } from "__generated__/MyCollectionArtworkListItem_artwork.graphql"
 import { MyCollectionQuery } from "__generated__/MyCollectionQuery.graphql"
 import { EventEmitter } from "events"
 import { ArtworkFilterNavigator, FilterModalMode } from "lib/Components/ArtworkFilter"
-import {
-  FilterData,
-  FilterDisplayName,
-  FilterParamName,
-} from "lib/Components/ArtworkFilter/ArtworkFilterHelpers"
 import {
   ArtworkFiltersStoreProvider,
   ArtworksFiltersStore,
@@ -24,20 +19,23 @@ import { StickyTabPageScrollView } from "lib/Components/StickyTabPage/StickyTabP
 import { useToast } from "lib/Components/Toast/toastHook"
 import { navigate, popToRoot } from "lib/navigation/navigate"
 import { defaultEnvironment } from "lib/relay/createEnvironment"
-import { useDevToggle } from "lib/store/GlobalStore"
+import { useDevToggle, useFeatureFlag } from "lib/store/GlobalStore"
 import { extractNodes } from "lib/utils/extractNodes"
 import { PlaceholderBox, PlaceholderGrid, PlaceholderText } from "lib/utils/placeholders"
 import { renderWithPlaceholder } from "lib/utils/renderWithPlaceholder"
 import { ProvideScreenTrackingWithCohesionSchema } from "lib/utils/track"
 import { screen } from "lib/utils/track/helpers"
 import { useScreenDimensions } from "lib/utils/useScreenDimensions"
-import { filter, orderBy, uniqBy } from "lodash"
-import { DateTime } from "luxon"
 import { Banner, Button, Flex, Separator, Spacer, useSpace } from "palette"
 import React, { useContext, useEffect, useState } from "react"
-import { RefreshControl } from "react-native"
+import { LayoutAnimation, RefreshControl } from "react-native"
 import { createPaginationContainer, graphql, QueryRenderer, RelayPaginationProp } from "react-relay"
 import { useTracking } from "react-tracking"
+import { MyCollectionSearchBar } from "./Components/MyCollectionSearchBar"
+import {
+  localSortAndFilterArtworks,
+  useLocalArtworkFilter,
+} from "./utils/localArtworkSortAndFilter"
 import { addRandomMyCollectionArtwork } from "./utils/randomMyCollectionArtwork"
 
 const RefreshEvents = new EventEmitter()
@@ -59,234 +57,23 @@ const MyCollection: React.FC<{
   me: MyCollection_me
 }> = ({ relay, me }) => {
   const { trackEvent } = useTracking()
-  const [filterModalVisible, setFilterModalVisible] = useState(false)
-  const filtersCount = useSelectedFiltersCount()
 
+  const enableSearchBar = useFeatureFlag("AREnableMyCollectionSearchBar")
+  const showDevAddButton = useDevToggle("DTEasyMyCollectionArtworkCreation")
+
+  const [keywordFilter, setKeywordFilter] = useState("")
+  const [isSearchBarVisible, setIsSearchBarVisible] = useState(false)
+  const [isFilterModalVisible, setIsFilterModalVisible] = useState(false)
+
+  const filtersCount = useSelectedFiltersCount()
   const appliedFiltersState = ArtworksFiltersStore.useStoreState((state) => state.appliedFilters)
 
   const artworks = extractNodes(me?.myCollectionConnection)
-  const [isRefreshing, setIsRefreshing] = useState(false)
 
-  const setFilterType = ArtworksFiltersStore.useStoreActions((s) => s.setFilterTypeAction)
-  const setSortOptions = ArtworksFiltersStore.useStoreActions((s) => s.setSortOptions)
-  const setFilterOptions = ArtworksFiltersStore.useStoreActions((s) => s.setFilterOptions)
   const filterOptions = ArtworksFiltersStore.useStoreState((s) => s.filterOptions)
+  useLocalArtworkFilter(artworks)
 
-  useEffect(() => {
-    setFilterType("local")
-    setSortOptions([
-      {
-        paramName: FilterParamName.sort,
-        displayText: "Price Paid (High to Low)",
-        paramValue: "local-price-paid-high-low",
-        // tslint:disable-next-line: no-shadowed-variable
-        localSortAndFilter: (artworks) =>
-          orderBy(
-            artworks,
-            (a) => {
-              return a.pricePaid?.minor
-            },
-            "asc"
-          ),
-      },
-      {
-        paramName: FilterParamName.sort,
-        displayText: "Price Paid (Low to High)",
-        paramValue: "local-price-paid-low-high",
-        // tslint:disable-next-line: no-shadowed-variable
-        localSortAndFilter: (artworks) =>
-          orderBy(
-            artworks,
-            (a) => {
-              return a.pricePaid?.minor
-            },
-            "desc"
-          ),
-      },
-      {
-        paramName: FilterParamName.sort,
-        displayText: "Artwork Year (Ascending)",
-        paramValue: "local-year-old-new",
-        // tslint:disable-next-line: no-shadowed-variable
-        localSortAndFilter: (artworks) =>
-          orderBy(
-            artworks,
-            (a) => {
-              const date = DateTime.fromISO(a.date)
-              return date.isValid ? date.toMillis() : Number.POSITIVE_INFINITY
-            },
-            "asc"
-          ),
-      },
-      {
-        paramName: FilterParamName.sort,
-        displayText: "Artwork Year (Descending)",
-        paramValue: "local-year-new-old",
-        // tslint:disable-next-line: no-shadowed-variable
-        localSortAndFilter: (artworks) =>
-          orderBy(
-            artworks,
-            (a) => {
-              const date = DateTime.fromISO(a.date)
-              return date.isValid ? date.toMillis() : 0
-            },
-            "desc"
-          ),
-      },
-      {
-        paramName: FilterParamName.sort,
-        displayText: "Alphabetical by Artist (A to Z)",
-        paramValue: "local-alpha-a-z",
-        // tslint:disable-next-line: no-shadowed-variable
-        localSortAndFilter: (artworks) => orderBy(artworks, (a) => a.artistNames, "asc"),
-      },
-      {
-        paramName: FilterParamName.sort,
-        displayText: "Alphabetical by Artist (Z to A)",
-        paramValue: "local-alpha-z-a",
-        // tslint:disable-next-line: no-shadowed-variable
-        localSortAndFilter: (artworks) => orderBy(artworks, (a) => a.artistNames, "desc"),
-      },
-    ])
-    setFilterOptions([
-      {
-        displayText: FilterDisplayName.sort,
-        filterType: "sort",
-        ScreenComponent: "SortOptionsScreen",
-      },
-      {
-        displayText: FilterDisplayName.artistIDs,
-        filterType: "artistIDs",
-        ScreenComponent: "ArtistIDsOptionsScreen",
-        values: uniqBy(
-          artworks.map(
-            (a): FilterData => ({
-              displayText: a.artist?.name ?? "N/A",
-              paramName: FilterParamName.artistIDs,
-              paramValue: a.artist?.internalID ?? "",
-            })
-          ),
-          (m) => m.paramValue
-        ),
-        // tslint:disable-next-line: no-shadowed-variable
-        localSortAndFilter: (artworks, artistIDs: string[]) =>
-          filter(artworks, (a) => artistIDs.includes(a.artist.internalID)),
-      },
-      {
-        displayText: FilterDisplayName.attributionClass,
-        filterType: "attributionClass",
-        ScreenComponent: "AttributionClassOptionsScreen",
-        // tslint:disable-next-line: no-shadowed-variable
-        localSortAndFilter: (artworks, attributionClasses: string[]) => {
-          return filter(artworks, (a) => {
-            if (a.attributionClass && a.attributionClass.name) {
-              return attributionClasses.includes(a.attributionClass.name)
-            }
-            return false
-          })
-        },
-      },
-      {
-        displayText: FilterDisplayName.additionalGeneIDs,
-        filterType: "additionalGeneIDs",
-        ScreenComponent: "AdditionalGeneIDsOptionsScreen",
-        values: uniqBy(
-          artworks.map(
-            (a): FilterData => ({
-              displayText: a.medium ?? "N/A",
-              paramName: FilterParamName.additionalGeneIDs,
-              paramValue: a.medium ?? undefined,
-            })
-          ),
-          (m) => m.paramValue
-        ),
-        // tslint:disable-next-line: no-shadowed-variable
-        localSortAndFilter: (artworks, mediums: string[]) =>
-          filter(artworks, (a) => mediums.includes(a.medium)),
-      },
-      {
-        displayText: FilterDisplayName.priceRange,
-        filterType: "priceRange",
-        ScreenComponent: "PriceRangeOptionsScreen",
-        // tslint:disable-next-line: no-shadowed-variable
-        localSortAndFilter: (artworks, priceRange: string) => {
-          const splitRange = priceRange.split("-")
-          const lowerBoundStr = splitRange[0]
-          const upperBoundStr = splitRange[1]
-
-          let lowerBound = 0
-          let upperBound = Number.POSITIVE_INFINITY
-
-          const parsedLower = parseInt(lowerBoundStr, 10)
-          const parsedUpper = parseInt(upperBoundStr, 10)
-
-          if (!isNaN(parsedLower)) {
-            lowerBound = parsedLower
-          }
-
-          if (!isNaN(parsedUpper)) {
-            upperBound = parsedUpper
-          }
-
-          return filter(artworks, (a) => {
-            if (isNaN(a.pricePaid?.minor)) {
-              return false
-            }
-            const pricePaid = a.pricePaid?.minor / 100
-            return pricePaid >= lowerBound && pricePaid <= upperBound
-          })
-        },
-      },
-      {
-        displayText: FilterDisplayName.sizes,
-        filterType: "sizes",
-        ScreenComponent: "SizesOptionsScreen",
-
-        localSortAndFilter: (
-          // tslint:disable-next-line: no-shadowed-variable
-          artworks,
-          sizeParams: {
-            paramName: FilterParamName.width | FilterParamName.height | FilterParamName.sizes
-            paramValue: string
-          }
-        ) => {
-          if (sizeParams.paramName === "sizes") {
-            return filter(artworks, (a) => {
-              const size: string = a.sizeBucket
-              return sizeParams.paramValue.includes(size.toUpperCase())
-            })
-          } else {
-            const splitRange = sizeParams.paramValue.split("-")
-            const lowerBoundStr = splitRange[0]
-            const upperBoundStr = splitRange[1]
-
-            let lowerBound = 0
-            let upperBound = Number.POSITIVE_INFINITY
-
-            const parsedLower = parseInt(lowerBoundStr, 10)
-            const parsedUpper = parseInt(upperBoundStr, 10)
-
-            if (!isNaN(parsedLower)) {
-              lowerBound = parsedLower
-            }
-
-            if (!isNaN(parsedUpper)) {
-              upperBound = parsedUpper
-            }
-
-            return filter(artworks, (a) => {
-              const targetMetric = sizeParams.paramName === "width" ? a.width : a.height
-              if (isNaN(targetMetric)) {
-                return false
-              }
-              return targetMetric >= lowerBound && targetMetric <= upperBound
-            })
-          }
-        },
-      },
-    ])
-  }, [])
-
+  const [isRefreshing, setIsRefreshing] = useState(false)
   useEffect(() => {
     RefreshEvents.addListener(REFRESH_KEY, refetch)
     return () => {
@@ -310,8 +97,6 @@ const MyCollection: React.FC<{
   const space = useSpace()
   const toast = useToast()
 
-  const showDevAddButton = useDevToggle("DTEasyMyCollectionArtworkCreation")
-
   useEffect(() => {
     if (artworks.length) {
       hasBeenShownBanner().then((hasSeenBanner) => {
@@ -321,7 +106,7 @@ const MyCollection: React.FC<{
           <Flex>
             <ArtworksFilterHeader
               selectedFiltersCount={filtersCount}
-              onFilterPress={() => setFilterModalVisible(true)}
+              onFilterPress={() => setIsFilterModalVisible(true)}
             >
               <Button
                 data-test-id="add-artwork-button-non-zero-state"
@@ -341,6 +126,9 @@ const MyCollection: React.FC<{
                 Add Works
               </Button>
             </ArtworksFilterHeader>
+            {!!enableSearchBar && !!isSearchBarVisible && (
+              <MyCollectionSearchBar onChangeText={setKeywordFilter} />
+            )}
             {!!showNewWorksBanner && (
               <Banner
                 title="Your collection is growing"
@@ -358,7 +146,7 @@ const MyCollection: React.FC<{
       // remove already set JSX
       setJSX(null)
     }
-  }, [artworks.length, filtersCount])
+  }, [artworks.length, filtersCount, isSearchBarVisible])
 
   return (
     <ProvideScreenTrackingWithCohesionSchema
@@ -367,38 +155,26 @@ const MyCollection: React.FC<{
       })}
     >
       <ArtworkFilterNavigator
-        visible={filterModalVisible}
+        visible={isFilterModalVisible}
         mode={FilterModalMode.Custom}
-        closeModal={() => setFilterModalVisible(false)}
-        exitModal={() => setFilterModalVisible(false)}
+        closeModal={() => setIsFilterModalVisible(false)}
+        exitModal={() => setIsFilterModalVisible(false)}
       />
 
       <StickyTabPageScrollView
         contentContainerStyle={{ paddingBottom: space(2) }}
         refreshControl={<RefreshControl refreshing={isRefreshing} onRefresh={refetch} />}
+        onScrollBeginDrag={() => {
+          if (isSearchBarVisible) {
+            return
+          }
+
+          LayoutAnimation.configureNext(LayoutAnimation.Presets.spring)
+          setIsSearchBarVisible(true)
+        }}
       >
         {artworks.length === 0 ? (
-          <ZeroState
-            title="Your art collection in your pocket."
-            subtitle="Keep track of your collection all in one place and get market insights"
-            callToAction={
-              <Button
-                testID="add-artwork-button-zero-state"
-                onPress={() => {
-                  trackEvent(tracks.addCollectedArtwork())
-                  navigate("my-collection/artworks/new", {
-                    passProps: {
-                      mode: "add",
-                      onSuccess: popToRoot,
-                    },
-                  })
-                }}
-                block
-              >
-                Add artwork
-              </Button>
-            }
-          />
+          <MyCollectionZeroState />
         ) : (
           <Flex pt={1}>
             <InfiniteScrollMyCollectionArtworksGridContainer
@@ -406,57 +182,14 @@ const MyCollection: React.FC<{
               hasMore={relay.hasMore}
               loadMore={relay.loadMore}
               // tslint:disable-next-line: no-shadowed-variable
-              localSortAndFilterArtworks={(artworks: MyCollectionArtworkListItem_artwork[]) => {
-                let processedArtworks = artworks
-
-                const filtering = uniqBy(
-                  appliedFiltersState.filter((x) => x.paramName !== FilterParamName.sort),
-                  (f) => f.paramName
+              localSortAndFilterArtworks={(artworks: MyCollectionArtworkEdge[]) =>
+                localSortAndFilterArtworks(
+                  artworks,
+                  appliedFiltersState,
+                  filterOptions,
+                  keywordFilter
                 )
-
-                // custom size filters come back with a different type, consolidate to one
-                const sizeFilterTypes = [
-                  FilterParamName.width,
-                  FilterParamName.height,
-                  FilterParamName.sizes,
-                ]
-
-                // tslint:disable-next-line: no-shadowed-variable
-                filtering.forEach((filter) => {
-                  if (sizeFilterTypes.includes(filter.paramName)) {
-                    /*
-                     * Custom handling for size filter
-                     * 2 flavors:
-                     * a sizeRange representing either a width or height restriction OR
-                     * 1 or more size bucket names which should be matched against artwork values
-                     * pass the paramName so we can distinguish how to handle in the step
-                     */
-                    const sizeFilterParamName = FilterParamName.sizes
-                    const sizeFilterStep = (filterOptions ?? []).find(
-                      (f) => f.filterType === sizeFilterParamName
-                    )!.localSortAndFilter!
-                    processedArtworks = sizeFilterStep(processedArtworks, {
-                      paramValue: filter.paramValue,
-                      paramName: filter.paramName,
-                    })
-                  } else {
-                    const filterStep = (filterOptions ?? []).find(
-                      (f) => f.filterType === filter.paramName
-                    )!.localSortAndFilter!
-                    processedArtworks = filterStep(processedArtworks, filter.paramValue)
-                  }
-                })
-
-                const sorting = appliedFiltersState.filter(
-                  (x) => x.paramName === FilterParamName.sort
-                )
-                if (sorting.length > 0) {
-                  const sortStep = sorting[0].localSortAndFilter!
-                  processedArtworks = sortStep(processedArtworks)
-                }
-
-                return processedArtworks
-              }}
+              }
             />
           </Flex>
         )}
@@ -493,6 +226,7 @@ export const MyCollectionContainer = createPaginationContainer(
             node {
               id
               medium
+              title
               pricePaid {
                 minor
               }
@@ -607,6 +341,38 @@ export const MyCollectionPlaceholder: React.FC<{}> = () => {
   )
 }
 
+const MyCollectionZeroState: React.FC = () => {
+  const { trackEvent } = useTracking()
+
+  return (
+    <ZeroState
+      title="Your art collection in your pocket."
+      subtitle="Keep track of your collection all in one place and get market insights"
+      callToAction={
+        <Button
+          testID="add-artwork-button-zero-state"
+          onPress={() => {
+            trackEvent(tracks.addCollectedArtwork())
+            navigate("my-collection/artworks/new", {
+              passProps: {
+                mode: "add",
+                onSuccess: popToRoot,
+              },
+            })
+          }}
+          block
+        >
+          Add artwork
+        </Button>
+      }
+    />
+  )
+}
+
 const tracks = {
   addCollectedArtwork,
 }
+
+export type MyCollectionArtworkEdge = NonNullable<
+  NonNullable<InfiniteScrollArtworksGrid_myCollectionConnection["edges"]>[0]
+>["node"]
